zeroize: Replace 'secure_zero_memory' with 'Zeroize'
This commit makes the `secure_zero_memory()` function `pub(crate)` and
uses it to implement the `Zeroize` trait for `AsMut<[u8]>`, which is now
the only API provided by this crate.

It might be worth adding back for certain cases(?), but this crate's
original focus was having a very simple API, and with both there's
confusion as to which to use.

Removing `secure_zero_memory()` from the public API makes it *very*
clear: `Zeroize` is the API to use.