subtle-encoding: Fix encode/decode in release builds
The call to invoke the actual encoding/decoding functions was
unintentionally gated on a `debug_assert_eq!`, and therefore was never
invoked in release builds, resulting in all zeroes for encode/decode
operations.

This change moves these invocations outside of the `debug_assert_eq!`
statements so they are invoked in release builds.