zeroize: Remove 'nightly' feature
Zeroize previously provided a 'nightly' feature which used a slightly
more complex implementation in order to facilitate a micro-optimization.

However, the extra complexity was never justified, nor does it represent
the desired path forward on nightly (which would be to use atomic writes
instead of volatile writes, when they become available but aren't yet
stabilized).

It also seems some impls are missing, particularly for arrays, in which
case the special-case nightly impls become more cumbersome.

This commit rips out the entire nightly feature and ensures the crate
behaves the same way everywhere regardless of Rust channel.