hkd32: Merge code from `tiny-bip39`
All of the `bip39` code is high-sensitivity in that it deals with master
derivation passwords for secret keys. Unfortunately, the `bip39` crate
seems abandoned, which is not a good state for such a high-sensitivity
crate to be in.

This commit merges in the code from `tiny-bip39` as of the following
git commit bumping the version to v0.6.2:

> Version bump

https://github.com/maciejhirsz/tiny-bip39/commit/dfcb9c959fc4f411736cc9536af18c60b486d60f

This commit makes no other changes to the incorporated files, and they
have not been "wired up" to the rest of `hkd32` yet.