secrecy: Replace `bytes-serde` feature with custom deserializer
It's tricky with Rust's feature system to represent the combination of
features in such a way that it can conditionally activate another
feature of one of the crates when that crate is itself a feature.

This instead handles feature unification for `bytes` and `serde` through
a not-too-terribly-compilicated custom deserializer such that merely
having both features active is enough to receive the impl.