Ensure Option is None when zeroized (#649)
When Option is zeroized, set self to None

Say you had a `Vec<Option<T>> where T: Zeroize`, and then you called
`zeroize` on the `Vec`. The bit pattern left in memory would have
zeroized T, but the pattern of `Some` and `None` within the `Vec` would
remain.

This change prevents such a leak by setting `self` to the default in a
manner similar to how `DefaultIsZeroes` implements it.