iqhttp: add `HttpsClient::{add_header, headers_mut}` (#981)
Adds support for configuring HTTP headers to send as part of the
request.

Uses `hyper::HeaderMap` as the internal data structure for headers,
making them yet another thin wrapper around `hyper`.

This functionality existed somewhat before, but as a crate-internal
feature only. This commit exposes it as a first-class feature.